Changed old reference to ConMan into SLURM. Changed backup controller to "lx01" in slurm.conf2 Changed remaing malloc/free functions to xmalloc/xfree.

+++ trunk/src/common/log.c
@@ -157,7 +157,7 @@ static char *vxstrfmt(const char *fmt, va_list ap)
 		} else {        /* *p == '%' */
 
 			/* take difference from fmt to just before `%' */
-			len = (size_t) ((int)(p) - (int)fmt);
+			len = (size_t) ((long)(p) - (long)fmt);
 
 			/* append from fmt to p into buf if there's 
 			 * anythere there
